+++ CHANGELOG.md
@@ -12,3 +12,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 - Queue order function now takes into account potential victims, resulting in better reclaim scenarios.
+
+### Fixes
+- Fixed preempt/reclaim of elastic workloads only taking one pod.